Add significant portion of model, cleanup
Add a stub BoardView and PlayerLevelView. Add a PlayerLoadLevelCtrl
which loads
a level with sane defaults.

Implement Board, Cell, CellType, Level, Move, Point, Rules, Tile, and
Variation.

Add a WeightedRandom utility class for selecting a new Tile based on the
number
weights.

Remove the Game and Builder classes to correctly restructure the project
for
EBC.

Signed-off-by: Eli Skeggs <esskeggs@wpi.edu>